AG-11673 TC3 Rename datumUtils.ts to datumUtil.ts
This matches the naming pattern of other files in this directory.
